chore(opapi): move function implementation to generator.ts
