+++ opapi/src/section-types-generator/index.ts
@@ -1,1 +1,3 @@
-export { generateTypesBySection } from './section-types-generator'
+export type { DefaultState } from './section-types-generator.types'
+export { executeSectionParsers, executeOperationParsers, composeFilesFromBlocks } from './section-types-generator'
+export { sectionParsers, operationParsers } from './constants'
